Data file now generated by script
Previously the data file containg files not in the target Moodle version
was generated by steps listed in README.  Now the file is generated by a
Bash script.  The file has been renamed and is formatted with leading
'/' so that the output in both tables in consistent.  The file also now
contains the files checked by upgrade_stale_php_files_present() so
softac_check() has been updated to remove these when it reads the file.